Get rid of the Sphinx duplicate label warnings.
Labels and reference in Sphinx works like below:
1. Define a unique label across the whole document.
2. Use :ref: to link to this internal label.

The problem is, multiple "table" labels are defined with the same name: "table" or "hardware-table" in different rst include files.
When Sphinx decides which are the targets, it searches the whole documents and finds a lot of labels, then it can potentially link to the wrong label.

For example, it might link the "check the table below" of IMX6Q to the "hardware-table" label of Pi.

Since the tables are mostly right under the paragraph text, we may as well remove these labels.

Not all labels and :ref: should be removed, e.g.: those linking to OS downloads page, they work fine, because these labels do NOT have duplicated label names.

I have checked there is no dangling :ref:s in the document, if it did have one, the compiler should complain and give a warning -- but I am not seeing any, so I assume we are safe here.