Fix a bug where non-existing 3D model link appear in the doc.
After 404 is handle by a 302 on the doc's hosting provider, the testing of whether a 3D model exists is not working.
E.g. a product has PPC 3D model but not EPC 3D model, this 302 will tell it a EPC 3D model exist.
After the fix, the non-existing EPC 3D model link will not appear anymore.